+++ WalletWasabi/Wallets/Wallet.cs
@@ -95,8 +95,10 @@ namespace WalletWasabi.Wallets
 
 		public bool IsLoggedIn { get; private set; }
 
+		private Kitchen Kitchen { get; } = new();
+
 		public void Login() => IsLoggedIn = true;
-		
+
 		public void Logout() => IsLoggedIn = false;
 
 		public void RegisterServices(
@@ -120,7 +122,7 @@ namespace WalletWasabi.Wallets
 				ServiceConfiguration = Guard.NotNull(nameof(serviceConfiguration), serviceConfiguration);
 				FeeProvider = Guard.NotNull(nameof(feeProvider), feeProvider);
 
-				ChaumianClient = new CoinJoinClient(Synchronizer, Network, KeyManager);
+				ChaumianClient = new CoinJoinClient(Synchronizer, Network, KeyManager, Kitchen);
 
 				TransactionProcessor = new TransactionProcessor(BitcoinStore.TransactionStore, KeyManager, ServiceConfiguration.DustThreshold, ServiceConfiguration.PrivacyLevelStrong);
 				Coins = TransactionProcessor.Coins;
@@ -294,7 +296,7 @@ namespace WalletWasabi.Wallets
 						foreach (var newCoin in newCoins)
 						{
 							// If it's being mixed and anonset is not sufficient, then queue it.
-							if (!newCoin.IsSpent() && ChaumianClient.HasIngredients
+							if (!newCoin.IsSpent() && Kitchen.HasIngredients
 								&& newCoin.HdPubKey.AnonymitySet < ServiceConfiguration.GetMixUntilAnonymitySetValue())
 							{
 								coinsToQueue.Add(newCoin);
